Add a content part for references
